requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Environment variables
python-dotenv==1.0.0

# OpenAI (for future phases)
openai>=1.12.0

# Database (for future phases)
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
asyncpg==0.29.0

# PDF Processing (for future phases)
PyPDF2==3.0.1
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0

# Vector operations (for future phases)
numpy==1.26.2
pgvector==0.2.3

# Utilities
requests==2.31.0
aiofiles==23.2.1
pydantic==2.5.0
pydantic-settings==2.1.0